refactor and get helpers
